refactor(rest): extract plugin/catalog listing routes into routes/plugins.py
First route group out of the create_app closure, setting the pattern for
the rest (documented in routes/__init__.py): a prefix-less APIRouter that
create_app includes where the routes used to sit (route order matters —
the fixed-segment procedural-models/<catalog> paths must precede
procedural-models/{model_id}), with closure state reached through an
explicit RestContext (settings/storage/queue) on app.state via Depends,
not the closure.

Helpers the group needed are lifted to module level in routes/deps.py —
parse_scope / resolve_project_scope / scope_from_path / scope_from_header
and live_worker_specs(queue, ...) (+ _merge_spec) — and the plugin admin
gate (PLUGIN_JOB_ADMIN_SETTING, plugin_ids_gated_by_config) moves with its
group. create_app keeps the old underscore names bound to them, so the
~120 closure call sites are untouched and every response is unchanged.
