+++ src/ada/comms/rest/app.py
@@ -34,7 +34,6 @@ from . import auth as auth_module
 from . import db as db_module
 from . import failure_capture, local_jobs, pending_uploads
 from .auth import User
-from .catalog import merge_catalog_specs, overlay_catalog_rows, sort_by_name
 from .config import Settings, load_settings
 from .converter import (
     LEGACY_CONVERT_EXTS,
@@ -53,11 +52,7 @@ from .converter import (
     supported_targets_for,
 )
 from .handlers import dispatch
-from .plugin_registry import (
-    discover_local_plugins,
-    locally_registered_spec,
-    locally_registered_specs,
-)
+from .plugin_registry import discover_local_plugins, locally_registered_spec
 from .procedural import (
     procedural_build_job_key,
     procedural_detail_job_key,
@@ -69,6 +64,17 @@ from .procedural import (
 )
 from .qualification import CAPABILITY_REQUIREMENTS_KEY
 from .queue import JobQueue, capability_token
+from .routes.deps import (  # noqa: F401 — _merge_spec re-exported for tests/importers of the old name
+    RestContext,
+    _merge_spec,
+    live_worker_specs,
+    parse_scope,
+    resolve_project_scope,
+    scope_from_header,
+    scope_from_path,
+)
+from .routes.plugins import plugin_ids_gated_by_config
+from .routes.plugins import router as plugins_router
 from .scope import Scope
 from .scope import can_access as scope_can_access
 from .storage import Storage
@@ -181,54 +187,6 @@ def _pending_upload_detail(key: str, pending: pending_uploads.PendingUpload) -> 
 #: The `app_settings` key an admin edits. Mirrored into the KV meta keyspace
 #: under `CAPABILITY_REQUIREMENTS_KEY` so workers without a database can read it.
 CAPABILITY_REQUIREMENTS_SETTING = "capability_requirements"
-
-
-def _merge_spec(base: dict, other: dict) -> None:
-    """Fold a second worker's advertisement of the same slug into ``base``.
-
-    Only the keys named in ``union_fields`` are combined, and only where both
-    sides hold lists; everything else keeps the value the first worker supplied
-    (see :func:`_live_worker_specs` for why "first" is well-defined). Order is
-    preserved and duplicates dropped, so the result reads like one list somebody
-    wrote rather than a concatenation.
-
-    ``union_fields`` is itself unioned. That is what makes a rolling upgrade
-    work: while half the pool runs a build that declares the key and half does
-    not, the half that does still gets its fields merged instead of the
-    behaviour flipping on whichever worker sorted first.
-    """
-
-    def _union(dst: list, src: list) -> list:
-        seen = {json.dumps(v, sort_keys=True) if isinstance(v, (dict, list)) else v for v in dst}
-        for v in src:
-            marker = json.dumps(v, sort_keys=True) if isinstance(v, (dict, list)) else v
-            if marker not in seen:
-                seen.add(marker)
-                dst.append(v)
-        return dst
-
-    fields = base.get("union_fields")
-    fields = list(fields) if isinstance(fields, list) else []
-    incoming = other.get("union_fields")
-    if isinstance(incoming, list):
-        fields = _union(fields, [f for f in incoming if isinstance(f, str)])
-        base["union_fields"] = fields
-
-    for key in fields:
-        if not isinstance(key, str) or key == "union_fields":
-            continue
-        add = other.get(key)
-        if not isinstance(add, list):
-            continue
-        have = base.get(key)
-        if not isinstance(have, list):
-            # The first worker did not carry the key at all (older build, or it
-            # genuinely has nothing to contribute). Start from what this one
-            # has rather than dropping it.
-            base[key] = list(add)
-        else:
-            _union(have, add)
-
 
 _ADAPY_VERSION: str | None = None
 
@@ -460,6 +418,9 @@ def create_app(settings: Settings | None = None) -> FastAPI:
         lifespan=lifespan,
     )
     auth_module.install(app, settings.auth)
+    # Explicit per-app services for extracted routers (routes/*.py) — what
+    # they may reach instead of this closure. See routes/__init__.py.
+    app.state.rest = RestContext(settings=settings, storage=storage, queue=queue)
 
     @app.get("/healthz")
     async def healthz() -> Response:
@@ -730,91 +691,13 @@ def create_app(settings: Settings | None = None) -> FastAPI:
 
     # ── Scope helpers ────────────────────────────────────────────────
     #
-    # Scope wire format: a single path segment / header value, one of
-    #   shared          — the shared bucket (any auth user)
-    #   user:me         — the caller's personal scope (resolved server-
-    #                     side to user.sub so URLs are user-agnostic)
-    #   project:<id>    — a project the caller is a member of
-    #
-    # Membership and project existence are checked against the DB; with
-    # no DB, project scopes are categorically inaccessible.
-
-    def _parse_scope(s: str, user: User) -> Scope:
-        if s == "shared":
-            return Scope.shared()
-        if s == "user:me":
-            return Scope.user(user.sub)
-        if s.startswith("user:"):
-            # Naming another user explicitly is intentionally not
-            # allowed; admins use phase-3 admin endpoints instead.
-            raise HTTPException(
-                status_code=400,
-                detail="use 'user:me' for personal scope",
-            )
-        if s.startswith("project:"):
-            pid = s[len("project:") :].strip()
-            if not pid:
-                raise HTTPException(status_code=400, detail="missing project id")
-            return Scope.project(pid)
-        if s.startswith("corpus:"):
-            slug = s[len("corpus:") :].strip()
-            if not slug:
-                raise HTTPException(status_code=400, detail="missing corpus slug")
-            # Admin-only gate fires in scope_can_access; here we just
-            # parse. Non-admin requests hit a 403 at the access check.
-            return Scope.corpus(slug)
-        raise HTTPException(status_code=400, detail=f"invalid scope {s!r}")
-
-    async def _resolve_project_scope(pool, scope: Scope) -> Scope:
-        """Resolve ``project:<slug>`` to ``project:<uuid>`` against the DB.
-
-        ``_parse_scope`` doesn't know whether the id segment is a UUID or
-        a slug — it just hands the raw string through. UUID-shaped ids
-        pass through unchanged; non-UUID strings get looked up against
-        ``projects.slug`` so callers can use the friendlier form in
-        URLs and config files. Without a DB, slug lookup is impossible
-        and ``can_access`` will reject regardless, so we leave the scope
-        as-is.
-        """
-        if scope.kind != "project" or scope.id is None or pool is None:
-            return scope
-        import uuid as _uuid
-
-        try:
-            _uuid.UUID(scope.id)
-            return scope
-        except (ValueError, AttributeError, TypeError):
-            pass
-        resolved = await db_module.project_id_from_slug(pool, scope.id)
-        if resolved is None:
-            # Don't leak existence: same status as the membership check
-            # below would have produced for a non-member of an unknown
-            # project.
-            raise HTTPException(status_code=403, detail="forbidden")
-        return Scope.project(resolved)
-
-    async def _scope_from_path(
-        scope: str,
-        request: Request,
-        user: User = Depends(auth_module.current_user),
-    ) -> Scope:
-        s = _parse_scope(scope, user)
-        pool = getattr(request.app.state, "db_pool", None)
-        s = await _resolve_project_scope(pool, s)
-        if not await scope_can_access(user, s, pool):
-            raise HTTPException(status_code=403, detail="forbidden")
-        return s
-
-    async def _scope_from_header(
-        request: Request,
-        user: User = Depends(auth_module.current_user),
-    ) -> Scope:
-        s = _parse_scope(request.headers.get("X-Scope", "shared"), user)
-        pool = getattr(request.app.state, "db_pool", None)
-        s = await _resolve_project_scope(pool, s)
-        if not await scope_can_access(user, s, pool):
-            raise HTTPException(status_code=403, detail="forbidden")
-        return s
+    # The scope wire format + resolvers live in routes/deps.py (module-level,
+    # so extracted routers can ``Depends`` on them); the closure keeps the old
+    # names for the routes still defined in here.
+    _parse_scope = parse_scope
+    _resolve_project_scope = resolve_project_scope
+    _scope_from_path = scope_from_path
+    _scope_from_header = scope_from_header
 
     async def _audit(
         request: Request | None,
@@ -3543,143 +3426,11 @@ def create_app(settings: Settings | None = None) -> FastAPI:
         return JSONResponse(row)
 
     async def _live_worker_specs(field: str, fallback_field: str | None = None) -> dict[str, dict]:
-        """Catalog-shaped specs advertised by non-stale workers, keyed by slug.
-        Falls back to a bare slug-list field for older workers that advertise
-        only names, synthesizing a minimal spec.
+        # Closure-bound shorthand for routes still defined in here; see
+        # ``routes.deps.live_worker_specs`` for the semantics.
+        return await live_worker_specs(queue, field, fallback_field)
 
-        WHEN SEVERAL WORKERS ADVERTISE ONE SLUG, two rules apply:
-
-        * **Deterministic, not last-writer-wins.** Workers are visited in
-          worker-id order and the first spec for a slug supplies the scalars.
-          Previously this followed KV listing order, so with two workers on one
-          plugin the advertised version and capability could differ between two
-          consecutive requests for no visible reason.
-        * **Declared list fields are unioned.** A spec may name keys in
-          ``union_fields``; those are combined across every worker advertising
-          the slug instead of one worker's copy winning. That is what lets a
-          sharded pool say what it collectively covers — several workers on the
-          same plugin, each serving a different project, produce one spec
-          listing every project that is online.
-
-        Only the named keys are unioned. A blanket "merge every list" would
-        quietly combine things that are per-worker facts rather than collective
-        ones (a worker's own conversions, its own extension allowlist), and
-        produce a spec describing a worker that does not exist.
-        """
-        import time as _time
-
-        out: dict[str, dict] = {}
-        if not queue.enabled:
-            return out
-        now = _time.time()
-        # Sorted so the winner is stable across requests. `worker_id` is always
-        # present — list_workers derives it from the KV key.
-        workers = sorted(await queue.list_workers(), key=lambda w: str(w.get("worker_id") or ""))
-        for w in workers:
-            hb = w.get("last_heartbeat")
-            if not (isinstance(hb, (int, float)) and (now - hb) <= queue.WORKER_STALE_AFTER_S):
-                continue
-            # What this worker declined to serve, and why. A spec whose
-            # capability is withheld is carried through as UNAVAILABLE rather
-            # than dropped: a withheld capability and an absent one look
-            # identical to every consumer otherwise, and the consumer then
-            # tells the operator to start a worker that is already running.
-            # See deploy/worker-trust.md §4.
-            withheld = {
-                str(x.get("capability", "")).strip().lower(): str(x.get("reason") or "")
-                for x in (w.get("withheld") or [])
-                if isinstance(x, dict) and x.get("capability")
-            }
-            specs = w.get(field)
-            if isinstance(specs, list):
-                for s in specs:
-                    if isinstance(s, dict) and isinstance(s.get("slug"), str) and s["slug"]:
-                        slug = s["slug"]
-                        cap = str(s.get("worker_capability") or "").strip().lower()
-                        reason = withheld.get(cap) if cap else None
-                        entry = dict(s)
-                        if reason:
-                            entry["available"] = False
-                            entry["unavailable_reason"] = reason
-                        if slug not in out:
-                            out[slug] = entry
-                        elif reason is None and out[slug].get("available") is False:
-                            # A FIT worker overrides an unfit one's verdict: the
-                            # capability is available from somewhere, which is
-                            # what the caller actually needs to know. Order of
-                            # workers must not decide this.
-                            merged = dict(out[slug])
-                            merged.pop("available", None)
-                            merged.pop("unavailable_reason", None)
-                            out[slug] = merged
-                            _merge_spec(out[slug], entry)
-                        else:
-                            _merge_spec(out[slug], entry)
-            elif fallback_field:
-                bare = w.get(fallback_field)
-                if isinstance(bare, list):
-                    for slug in bare:
-                        if isinstance(slug, str) and slug and slug not in out:
-                            out[slug] = {"slug": slug, "name": slug.replace("_", " ").title()}
-        return out
-
-    # Plugin jobs an admin must be to enqueue, named by plugin id. Admin-only to
-    # write like every other non-`public.` setting, and deliberately NOT under
-    # `public.` — who may run a job is not a read window for every user.
-    #
-    # This exists because the plugin's own advertisement cannot be the only
-    # source. A spec is advertised BY THE WORKER (`_live_worker_specs`), so a
-    # worker running an older build that predates the flag advertises no flag,
-    # and a gate that reads only the advertisement would silently disappear —
-    # the deployment would look protected and be open, which is the failure
-    # this whole mechanism exists to prevent. Worker registry rows are also not
-    # currently unforgeable (see the `$KV.ada-viewer-jobs.>` note in adapy's
-    # worker-trust docs), so an advertisement is a statement of intent, not an
-    # authorization decision.
-    #
-    # The two sources are therefore OR-ed, never AND-ed: adding a source can
-    # only ever TIGHTEN. A stale worker cannot open a gate the deployment set,
-    # and a deployment can gate a plugin that never declared anything.
-    PLUGIN_JOB_ADMIN_SETTING = "admin.plugin_jobs.require_admin"
-
-    async def _plugin_ids_gated_by_config(pool) -> set[str] | None:
-        """Plugin ids the DEPLOYMENT says are admin-only.
-
-        Returns ``None`` for "the setting exists but could not be read", which
-        callers must treat as *every* plugin job requiring admin. Failing closed
-        on a malformed value is deliberate: the alternative is a typo quietly
-        removing a gate, and an over-tight gate announces itself immediately
-        while an absent one does not.
-        """
-        if pool is None:
-            return set()
-        try:
-            raw = await db_module.get_setting(pool, PLUGIN_JOB_ADMIN_SETTING)
-        except Exception:
-            # Could not ask. Same answer as a value we could not parse: gate
-            # everything. A gate must never become a 500 — that turns "the
-            # database hiccuped" into "the endpoint is broken" — and it must
-            # never resolve an error to "allowed", which would make an outage
-            # into a silent removal of the restriction.
-            logger.exception(
-                "api: could not read %s; treating every plugin job as admin-only",
-                PLUGIN_JOB_ADMIN_SETTING,
-            )
-            return None
-        if raw is None or not raw.strip():
-            return set()
-        try:
-            parsed = json.loads(raw)
-        except (ValueError, TypeError):
-            # Not JSON. Accept the shape a person types into a settings box
-            # rather than rejecting it, since rejecting means failing closed.
-            parts = [p.strip() for p in raw.replace(",", " ").split()]
-            return {p for p in parts if p} or None
-        if isinstance(parsed, str):
-            parsed = [parsed]
-        if not isinstance(parsed, list) or any(not isinstance(p, str) for p in parsed):
-            return None
-        return {p.strip() for p in parsed if p and p.strip()}
+    _plugin_ids_gated_by_config = plugin_ids_gated_by_config
 
     async def _plugin_job_requires_admin(plugin_id: str, pool, advertised: dict | None) -> bool:
         """Whether enqueuing ``plugin_id``'s job requires an admin.
@@ -3914,319 +3665,12 @@ def create_app(settings: Settings | None = None) -> FastAPI:
         value = await db_module.get_setting(pool, key)
         return JSONResponse({"key": key, "value": value})
 
-    @api.get("/plugins")
-    async def api_plugins(request: Request) -> JSONResponse:
-        """Backend plugins advertised to the viewer plugin system: the union of
-        the static built-ins (``builtin_plugin_specs`` — empty in core) and any a
-        live (non-stale) worker advertises via ``plugin_specs`` (its
-        ``ADA_WORKER_PRELOAD`` / ``ada.plugins`` entry point registered it with
-        ``register_plugin_backend``), keyed by ``slug`` and tagged ``origin``
-        ``code``/``db`` + ``online:true``. No hardcoded plugins: a backend plugin
-        appears only while a pool that provides it is online — the same
-        self-describing contract the procedural/detailing engines use. The
-        frontend build-time registry can seed off this so a runtime-only backend
-        plugin still surfaces."""
-        from .catalog import builtin_plugin_specs
-
-        # With no queue, plugin jobs run in THIS process (see `local_jobs`), so
-        # what it registered is online by definition — and the only source there
-        # is. Only then: behind a queue a job goes to a worker, and a spec this
-        # API happens to have imported says nothing about whether one is up.
-        by_slug = merge_catalog_specs(
-            builtin_plugin_specs(),
-            await _live_worker_specs("plugin_specs"),
-            project=lambda slug, spec, origin: {**spec, "slug": slug, "origin": origin, "online": True},
-            live_origin="db",
-            local_specs=None if queue.enabled else locally_registered_specs(),
-        )
-
-        # `requires_admin` is reported as the EFFECTIVE gate, not merely what a
-        # worker declared: a deployment can gate a plugin that declared nothing
-        # (see PLUGIN_JOB_ADMIN_SETTING). A UI that hid its button on the
-        # declaration alone would offer an action the API then refuses, which
-        # reads as a broken button rather than as a permission.
-        #
-        # This is an affordance, never the gate. The gate is in the POST.
-        pool = getattr(request.app.state, "db_pool", None)
-        gated = await _plugin_ids_gated_by_config(pool)
-        for slug, spec in by_slug.items():
-            spec["requires_admin"] = bool(spec.get("requires_admin")) or gated is None or slug in gated
-        return JSONResponse({"plugins": list(by_slug.values())})
-
-    @api.get("/scopes/{scope}/procedural-models/equipment-types")
-    async def api_procedural_equipment_types(
-        request: Request,
-        scope_obj: Scope = Depends(_scope_from_path),
-    ) -> JSONResponse:
-        """Equipment types for the cellbuilder's add-equipment dropdown: the
-        union of code-defined archetypes (advertised by live workers) and the
-        per-scope DB catalog, each tagged with its ``origin`` (``code`` or
-        ``catalog``) and its port list (name/direction/category plus local
-        position/direction_vector/colour — used by the viewer's missing-input
-        overlay and the port-glyph overlay). A slug present in both is shown as
-        ``catalog`` — the editable copy shadows the built-in."""
-
-        def _ports_of(doc: dict | None) -> list[dict]:
-            out = []
-            for p in (doc or {}).get("ports") or []:
-                if isinstance(p, dict) and p.get("name"):
-                    out.append(
-                        {
-                            "name": p["name"],
-                            "direction": p.get("direction", "INOUT"),
-                            "category": p.get("category", "process"),
-                            "position": p.get("position") or [0.0, 0.0, 0.0],
-                            "direction_vector": p.get("direction_vector") or [0.0, 0.0, 1.0],
-                            "color": p.get("color"),
-                        }
-                    )
-            return out
-
-        by_slug = merge_catalog_specs(
-            [],
-            await _live_worker_specs("procedural_equipment_specs", "procedural_equipment_types"),
-            project=lambda slug, spec, origin: {
-                "slug": slug,
-                "name": spec.get("name") or slug,
-                "origin": origin,
-                "ports": _ports_of(spec.get("doc")),
-                "has_cad": False,
-            },
-        )
-        pool = getattr(request.app.state, "db_pool", None)
-        if pool is not None:
-            # ``list_equipment_types`` returns summary rows only (no ``doc``), so
-            # fetch the docs separately to project each catalog type's ports.
-            docs_by_slug = await db_module.get_equipment_docs_by_scope(
-                pool, scope_kind=scope_obj.kind, scope_id=scope_obj.id
-            )
-            overlay_catalog_rows(
-                by_slug,
-                await db_module.list_equipment_types(pool, scope_kind=scope_obj.kind, scope_id=scope_obj.id),
-                lambda slug, t: {
-                    "slug": slug,
-                    "name": t.get("name") or slug,
-                    "origin": "catalog",
-                    "id": t["id"],
-                    "ports": _ports_of(docs_by_slug.get(slug)),
-                    # Whether a CAD asset is linked — drives the selected-object
-                    # "Show as CAD" toggle (which loads this type's preview GLB).
-                    "has_cad": bool(t.get("cad_key")),
-                },
-            )
-        return JSONResponse({"equipment_types": sort_by_name(by_slug.values())})
-
-    @api.get("/scopes/{scope}/procedural-models/system-types")
-    async def api_procedural_system_types(
-        request: Request,
-        scope_obj: Scope = Depends(_scope_from_path),
-    ) -> JSONResponse:
-        """System types for the cellbuilder's systems inspector: the union of
-        code-defined system kinds (piping/duct/cable/electrical — static, plus any
-        extra kinds advertised by live workers) and the per-scope DB
-        system-template catalog, each tagged with its ``origin`` and base kind."""
-        from .catalog import builtin_system_specs
-
-        def _project(slug: str, spec: dict, origin: str) -> dict:
-            doc = spec.get("doc") or {}
-            return {
-                "slug": slug,
-                "name": spec.get("name") or slug,
-                "origin": origin,
-                "type": doc.get("type", slug),
-                "medium": doc.get("medium"),
-                "voltage": doc.get("voltage"),
-            }
-
-        by_slug = merge_catalog_specs(
-            builtin_system_specs(),
-            await _live_worker_specs("procedural_system_specs", "procedural_system_types"),
-            project=_project,
-        )
-        pool = getattr(request.app.state, "db_pool", None)
-        if pool is not None:
-
-            def _project_row(slug: str, t: dict) -> dict:
-                doc = t.get("doc") or {}
-                return {
-                    "slug": slug,
-                    "name": t.get("name") or slug,
-                    "origin": "catalog",
-                    "id": t["id"],
-                    "type": doc.get("type", "piping"),
-                    "medium": doc.get("medium"),
-                    "voltage": doc.get("voltage"),
-                }
-
-            overlay_catalog_rows(
-                by_slug,
-                await db_module.list_system_templates(pool, scope_kind=scope_obj.kind, scope_id=scope_obj.id),
-                _project_row,
-            )
-        return JSONResponse({"system_types": sort_by_name(by_slug.values())})
-
-    @api.get("/scopes/{scope}/procedural-models/design-rulesets")
-    async def api_procedural_design_rulesets(
-        request: Request,
-        scope_obj: Scope = Depends(_scope_from_path),
-    ) -> JSONResponse:
-        """Named design rulesets for the cellbuilder's ruleset dropdown: the
-        built-in rulesets (static — ``standard``/``route_only``) plus any extra
-        advertised by live workers, each tagged ``origin`` ``code``. Selecting
-        one sets ``doc.design_rules``, which the compiler resolves to the routing/
-        penetration callables (``ada.topo_model.resolve_design_rules``)."""
-        from .catalog import builtin_design_rulesets
-
-        by_slug = merge_catalog_specs(
-            builtin_design_rulesets(),
-            await _live_worker_specs("procedural_design_rulesets"),
-            project=lambda slug, spec, origin: {
-                "slug": slug,
-                "name": spec.get("name") or slug,
-                "description": spec.get("description", ""),
-                "origin": origin,
-            },
-        )
-        return JSONResponse({"design_rulesets": sort_by_name(by_slug.values())})
-
-    @api.get("/scopes/{scope}/procedural-models/cell-types")
-    async def api_procedural_cell_types(
-        request: Request,
-        scope_obj: Scope = Depends(_scope_from_path),
-    ) -> JSONResponse:
-        """Space-cell types for the cellbuilder's ``+ Cell`` picker: the union of
-        the static built-in blueprints (``adapy-default``) and any advertised by
-        live workers (a capability engine registers its own via
-        ``register_procedural_cell_type``), each tagged ``origin`` ``code``. Each
-        carries a default box extent ``size`` ``(DX, DY, DZ)`` a freshly-placed
-        cell is seeded with, plus optional entity ``metadata``. No DB rows are
-        involved — like the design rulesets, so the dropdown is never empty for
-        want of a worker."""
-        from .catalog import builtin_cell_specs
-
-        by_slug = merge_catalog_specs(
-            builtin_cell_specs(),
-            await _live_worker_specs("procedural_cell_specs"),
-            project=lambda slug, spec, origin: {
-                "slug": slug,
-                "name": spec.get("name") or slug,
-                "origin": origin,
-                "size": spec.get("size") or [5.0, 5.0, 3.0],
-                "metadata": spec.get("metadata") or {},
-            },
-        )
-        return JSONResponse({"cell_types": sort_by_name(by_slug.values())})
-
-    @api.get("/scopes/{scope}/procedural-models/opening-types")
-    async def api_procedural_opening_types(
-        request: Request,
-        scope_obj: Scope = Depends(_scope_from_path),
-    ) -> JSONResponse:
-        """Opening types for the cellbuilder's ``+ Opening`` picker: the union of
-        the static built-in door/window/opening types (``adapy-default``) and any
-        advertised by live workers (via ``register_procedural_opening_type``), each
-        tagged ``origin`` ``code``. Each carries its ``subtype`` (``door``/
-        ``window``/``opening`` — the reinforcement framing the compiler frames
-        around the hole) and the
-        default box extent ``size`` ``(DX, DY, DZ)``. No DB rows are involved."""
-        from .catalog import builtin_opening_specs
-
-        by_slug = merge_catalog_specs(
-            builtin_opening_specs(),
-            await _live_worker_specs("procedural_opening_specs"),
-            project=lambda slug, spec, origin: {
-                "slug": slug,
-                "name": spec.get("name") or slug,
-                "origin": origin,
-                "subtype": spec.get("subtype") if spec.get("subtype") in ("door", "window", "opening") else "door",
-                "size": spec.get("size") or [1.0, 1.0, 2.0],
-            },
-        )
-        return JSONResponse({"opening_types": sort_by_name(by_slug.values())})
-
-    @api.get("/scopes/{scope}/procedural-models/blueprints")
-    async def api_procedural_blueprints(
-        request: Request,
-        engine: str = "adapy-default",
-        scope_obj: Scope = Depends(_scope_from_path),
-    ) -> JSONResponse:
-        """Structural blueprints for the cellbuilder's Blueprint dropdown, scoped
-        to the compile ``engine`` query param (default ``adapy-default``): the
-        union of that engine's static built-ins (``steel_stru``/``none`` for the
-        default engine) and any advertised by live workers for the SAME engine (a
-        capability engine registers its own via ``register_procedural_blueprint``),
-        each tagged ``origin`` ``code``. Selecting one sets the document's
-        ``blueprint_name``. The first entry is the engine's default; the list is
-        never empty — an engine advertising none falls back to an ``engine
-        default`` entry. No DB rows are involved."""
-        from .catalog import builtin_procedural_blueprint_specs
-
-        # Preserve authored order (built-ins first, the FIRST being the default),
-        # deduped by slug; live-worker extras append after. Engine-scoped: a live
-        # spec carries the engine it belongs to; keep only this engine's (a spec
-        # missing ``engine`` is treated as this one).
-        by_slug = merge_catalog_specs(
-            builtin_procedural_blueprint_specs(engine),
-            await _live_worker_specs("procedural_blueprint_specs"),
-            project=lambda slug, spec, origin: {
-                "slug": slug,
-                "name": spec.get("name") or slug,
-                "description": spec.get("description", ""),
-                "fields": spec.get("fields", []),
-                "origin": origin,
-            },
-            live_filter=lambda spec: spec.get("engine") in (None, engine),
-        )
-        blueprints = list(by_slug.values())
-        if not blueprints:
-            # An engine that advertised nothing (offline capability worker) still
-            # needs a non-empty dropdown so the compile can proceed.
-            blueprints = [
-                {
-                    "slug": "engine-default",
-                    "name": "Engine default",
-                    "description": "The engine's default blueprint.",
-                    "origin": "code",
-                }
-            ]
-        return JSONResponse({"blueprints": blueprints})
-
-    @api.get("/scopes/{scope}/procedural-models/detailing-engines")
-    async def api_procedural_detailing_engines(
-        request: Request,
-        scope_obj: Scope = Depends(_scope_from_path),
-    ) -> JSONResponse:
-        """Detailing engines for the Compile-settings "Detailing" dropdown: the
-        union of the static built-ins (``none`` + ``adapy-default``) and any an
-        external (out-of-process) engine a live (non-stale) capability worker
-        advertises via ``procedural_detailing_engine_specs``, each tagged ``origin``
-        ``code``/``db``. Selecting one is a COMPILE-time choice (not part of the
-        document); ``none`` (the default, first) passes no detailing -> structural-
-        only GLB. No hardcoded external engines: an external engine appears only
-        while its pool is online (modeled on the blueprints/design-rulesets
-        dropdowns)."""
-        from .catalog import builtin_detailing_engine_specs
-
-        # External engines are discovered ONLY from live capability workers (their
-        # ADA_WORKER_PRELOAD registers the engine, so the heartbeat advertises it).
-        # A live worker re-announcing a built-in keeps origin=code; a new
-        # (external) engine is worker/db-provided.
-        by_slug = merge_catalog_specs(
-            builtin_detailing_engine_specs(),
-            await _live_worker_specs("procedural_detailing_engine_specs"),
-            project=lambda slug, spec, origin: {
-                "slug": slug,
-                "name": spec.get("name") or slug,
-                "description": spec.get("description", ""),
-                "inprocess": bool(spec.get("inprocess", False)),
-                "worker_capability": spec.get("worker_capability"),
-                "joint_types": spec.get("joint_types", []),
-                "origin": origin,
-                "online": True,
-            },
-            live_origin="db",
-        )
-        return JSONResponse({"detailing_engines": list(by_slug.values())})
+    # Plugin + catalog listing routes live in routes/plugins.py (the pattern
+    # for splitting this closure — see routes/__init__.py). Included HERE, not
+    # after the closure, because route order matters: the fixed-segment
+    # ``procedural-models/<catalog>`` paths must register before the
+    # ``procedural-models/{model_id}`` routes below.
+    api.include_router(plugins_router)
 
     @api.post("/scopes/{scope}/procedural-models/equipment-types/sync", status_code=201)
     async def api_procedural_equipment_sync(
